requirements.txt
opencv-python
mediapipe==0.10.9
scikit-learn
pandas
joblib
absl-py
numpy<2.0.0
tensorflow==2.15.0
protobuf==3.20.3